+++ packages/styles/CHANGELOG.md
@@ -142,7 +142,7 @@
 ### Minor Changes
 
 - 0c556bb: [**`<mg-input-rich-text-editor>`**](http://core.pages.mgdis.fr/core-ui/core-ui/storybook/?path=/docs/molecules-inputs-mg-input-rich-text-editor--docs) new component
-- 0964f29: [**`<mg-illustrated-message>`**](http://core.pages.mgdis.fr/core-ui/core-ui/storybook/?path=/docs/molecules-mg-illustrated-message--docs) Update sizes ans spacings
+- 0964f29: [**`<mg-illustrated-message>`**](http://core.pages.mgdis.fr/core-ui/core-ui/storybook/?path=/docs/molecules-mg-illustrated-message--docs) Update sizes and spacings
 
 ### Patch Changes